Add test and fix bug in inet-address stream operators

+++ src/pieces/inet_address.cpp
@@ -154,7 +154,7 @@ DataStream& operator<<(DataStream& ds, const InetAddress& ia)
 
 DataStream& operator>>(DataStream& ds, InetAddress& ia)
 {
-    unsigned long addr;
+    unsigned int addr;
     ds >> addr;
 
     ia = InetAddress(addr);
